chore: Update tailwind.config.js and package.json
The tailwind.config.js file was updated to remove the '@tailwindcss/line-clamp' plugin and add the 'daisyui' plugin. This change improves the configuration of Tailwind CSS for the project.

The package.json file was updated to remove the '@tailwindcss/line-clamp' and '@vitalets/google-translate-api' dependencies. This change removes unused dependencies from the project.

Update tailwind.config.js and package.json