feat: Add sharp library for image processing
The package.json file was updated to include the "sharp" library as a dependency. This library provides image processing capabilities and will be used to optimize and manipulate images in the project. This change enhances the functionality of the application by enabling image-related operations.

Add sharp library for image processing