feat: Add loading component
This commit adds a new file `loading.jsx` which will be used to display a loading indicator in the application. This component will be imported and used in various parts of the application to indicate when data is being fetched or processed. This addition improves the user experience by providing visual feedback during loading operations.