feat: Add Suspense component to layout.js for lazy loading
This commit adds the Suspense component to the layout.js file in order to enable lazy loading of images. By wrapping the children component in the Suspense component, the images will be loaded asynchronously, improving the page load performance. This change enhances the user experience by reducing the initial load time of the page.