refactor: Update ModalAddProduct to remove disabled attribute on category option
This commit removes the disabled attribute on the category option in the ModalAddProduct component. Previously, the option was disabled, preventing users from selecting a category. By removing the disabled attribute, users can now select a category when adding a new product. This change improves the functionality and usability of the product creation process.