refactor: Remove disabled attribute on category option in ModalEditProduct
This commit removes the disabled attribute on the category option in the ModalEditProduct component. Previously, the option was disabled, preventing users from selecting a category. By removing the disabled attribute, users can now select a category when editing a product. This change improves the functionality and usability of the product editing process.